Time: 207 ms (40%), Space: 171.6 MB (30%) - LeetHub

+++ 3528-reach-end-of-array-with-max-score/3528-reach-end-of-array-with-max-score.cpp
@@ -2,13 +2,13 @@ class Solution {
 public:
     long long findMaximumScore(vector<int>& nums) {
 
-        int maxi=nums[0];
+        long long maxi=nums[0];
         long long ans=0;
 
         for(int i=1;i<nums.size();i++)
         {
             ans+=maxi;
-            maxi=max(maxi,nums[i]);
+            maxi=max(maxi,(long long)nums[i]);
         }
